+++ generators/src/common/constants.d.ts
@@ -1,0 +1,41 @@
+/**
+ * Constants used throughout the generator
+ * All constants have been migrated from JavaScript to TypeScript
+ * The following files can be removed after testing the migration:
+ * - generators/common/constants.js
+ * - generators/constants.js
+ */
+export declare const Constants: {
+    JAVA_VERSION: string;
+    SPRING_BOOT_VERSION: string;
+    SPRING_CLOUD_VERSION: string;
+    SPRING_CLOUD_AWS_VERSION: string;
+    SPRING_DEP_MNGMNT_VERSION: string;
+    DEFAULT_APP_VERSION: string;
+    SPRINGDOC_OPENAPI_VERSION: string;
+    COMMONS_IO_VERSION: string;
+    LOGSTASH_LOGBACK_ENCODER: number;
+    PALANTIR_JAVA_FORMAT_VERSION: string;
+    JACOCO_MIN_COVERAGE_REQUIRED: string;
+    GRADLE_GIT_PROPERTIES_PLUGIN_VERSION: string;
+    GRADLE_JACOCO_PLUGIN_VERSION: string;
+    GRADLE_SONAR_PLUGIN_VERSION: string;
+    GRADLE_OWASP_PLUGIN_VERSION: string;
+    GRADLE_BENMANES_VERSIONS_PLUGIN_VERSION: string;
+    GRADLE_SPOTLESS_PLUGIN_VERSION: string;
+    MAVEN_DEPENDENCY_CHECK_PLUGIN_VERSION: string;
+    MAVEN_PROPERTIES_PLUGIN_VERSION: string;
+    MAVEN_SONAR_PLUGIN_VERSION: string;
+    MAVEN_JACOCO_PLUGIN_VERSION: string;
+    MAVEN_SPOTLESS_PLUGIN_VERSION: string;
+    KEY_FLYWAY_MIGRATION_COUNTER: string;
+    KEY_LIQUIBASE_MIGRATION_COUNTER: string;
+    LOCALSTACK_IMAGE_VERSION: string;
+    LOCALSTACK_IMAGE: string;
+    POSTGRESQL_IMAGE_VERSION: string;
+    POSTGRESQL_IMAGE: string;
+    MARIADB_IMAGE_VERSION: string;
+    MARIADB_IMAGE: string;
+    MYSQL_IMAGE_VERSION: string;
+    MYSQL_IMAGE: string;
+};